Development: Don't import app-router / hot-reloader through next/link in application code
